Had to change double[] to float[] for java 5 compatibility.

+++ src/openmap/com/bbn/openmap/MapBean.java
@@ -1660,7 +1660,7 @@ public class MapBean extends JComponent implements ComponentListener, ContainerL
          */
         public Shape inverseTransform(Shape shape) {
 
-            double[] coords = new double[6];
+            float[] coords = new float[6];
             GeneralPath path = new GeneralPath(GeneralPath.WIND_EVEN_ODD);
 
             PathIterator pi = shape.getPathIterator(getInverseRotationTransform());
